PLS HELP ME
i have made the users able to set their own End Score. Everything works normally except when it asks for the players' names at the end. It seems that it has skipped gets() the player1's name (line 269)

p.s. i have the line 281 and line 302-304 on comment, incase you want to use it to go back to the first menu.